backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
google-generativeai>=0.7.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
